Reorganize base classes in new `base` submodule
