+++ src/inversion_ideas/__init__.py
@@ -2,35 +2,29 @@
 Ideas for inversion framework.
 """
 
+from . import base
 from . import utils
 from ._version import __version__
 from .conditions import ChiTarget
 from .constructors import create_inversion
 from .data_misfit import DataMisfit
-from .directives import Directive, MultiplierCooler
+from .directives import MultiplierCooler
 from .errors import ConvergenceWarning
 from .inversion import Inversion, InversionLog
-from .minimizer import ConjugateGradient, Minimizer
-from .objective_function import Combo, Objective, Scaled
+from .minimizers import ConjugateGradient
 from .regularization import TikhonovZero
-from .simulation import Simulation
 
 __all__ = [
     "ChiTarget",
-    "Combo",
     "ConjugateGradient",
     "ConvergenceWarning",
     "DataMisfit",
-    "Directive",
     "Inversion",
     "InversionLog",
-    "Minimizer",
     "MultiplierCooler",
-    "Objective",
-    "Scaled",
-    "Simulation",
     "TikhonovZero",
     "__version__",
+    "base",
     "create_inversion",
     "utils",
 ]
