+++ src/inversion_ideas/base/objective_function.py
