+++ src/inversion_ideas/data_misfit.py
@@ -7,7 +7,7 @@ import numpy.typing as npt
 from scipy.sparse import diags_array, sparray
 from scipy.sparse.linalg import LinearOperator, aslinearoperator
 
-from .objective_function import Objective
+from .base import Objective
 from .utils import cache_on_model
